+++ src/test_data/vector_fst/fst_003_linear_acceptor_1_label.rs
@@ -23,7 +23,7 @@ impl TestFst for LinearAcceptor1Label {
 
     fn get_fst() -> <Self as TestFst>::F {
         let labels = vec![32];
-        acceptor(labels.into_iter())
+        acceptor(&labels, IntegerWeight::ONE)
     }
 
     fn get_name() -> String {